% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hliTA.R
\docType{package}
\name{bitTA-package}
\alias{bitTA}
\alias{bitTA-package}
\title{TA 기능의 비트알 3rd party 패키지}
\description{
자연어 처리(Natural Language Processing), 텍스트 분석 모델 및
텍스트 분석을 위한 시각화와 도구 모음.
}
\details{
\tabular{ll}{
Package: \tab bitTA\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2022-05-17\cr
License: \tab Apache License (== 2.0)\cr
}
}
\author{
유충현
Maintainer: 유충현 <choonghyun.ryu@gmail.com>
}
