% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{collapse_noun}
\alias{collapse_noun}
\title{Extract Collapsed Noun}
\usage{
collapse_noun(
  doc,
  chunk = round(length(if (tibble::is_tibble(doc)) dplyr::pull(doc) else doc)/mc.cores),
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{doc}{character. 명사로만 구성될 문서를 만들 대상 텍스트 데이터}

\item{chunk}{integer. 병렬 작업 수행 시 처리 단위인 chunk}

\item{mc.cores}{integer. 병렬 작업 수행 시 사용할 코어의 개수}
}
\value{
character. 명사로만 구성된 텍스트
}
\description{
텍스트 문서에서 명사들을 토큰화한 후, 토큰화된 명사들을 공백으로
묶어서 텍스트 문서를 만듦
조회한다.
}
\examples{
\donttest{
collapse_noun(president_speech$doc[1:7])

# Collaboration with tidytext
library(dplyr)

nho_noun <- president_speech \%>\%
  filter(president \%in\% "노무현") \%>\%
  filter(str_detect(category, "^외교")) \%>\%
  mutate(doc_noun = collapse_noun(doc)) \%>\%
    tidytext::unnest_ngrams(
      noun_bigram,
      doc_noun,
      n = 2
   )
nho_noun
 
nho_noun$noun_bigram[1:5]
}
}
