% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{doc_quality}
\alias{doc_quality}
\title{Text Data Quality Correction Worker}
\usage{
doc_quality()
}
\value{
없음
}
\description{
정규표현식 기반의 텍스트 데이터  품질 보정 작업을 위한 Shiny 앱 호출
}
\examples{
\dontrun{
 library(bitTA)

 ## 텍스트 데이터  품질 보정 작업기(Shiny Web Application) 호출
 doc_quality()
}
}
\seealso{
\code{\link{morpho_mecab}}, \code{\link{morpho_hann}}
}
\author{
유충현
Maintainer: 유충현 <choonghyun.ryu@gmail.com>
}
