% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mecab.R
\name{install_mecab_ko}
\alias{install_mecab_ko}
\title{은전한닢 형태소분석기와 사전 설치}
\usage{
install_mecab_ko(x)
}
\description{
은전한닢 형태소분석기인 mecab-ko와 은전한닢 형태소분석기 사전인 
mecab-ko-dic을 사용자 환경에 설치한다.
}
\details{
Linux와 Mac은 소스를 가져다 컴파일하며, Windows는 바이너리를 가져다 복사한다.
}
\examples{
\dontrun{
install_mecab_ko()
}
}
\author{
유충현
Maintainer: 유충현 <choonghyun.ryu@gmail.com>
}
