% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{morpho_hann}
\alias{morpho_hann}
\title{Hannnanum 형태소 분석기 기반 토큰화}
\usage{
morpho_hann(
  x,
  type = c("morpheme", "noun", "noun2", "verb", "adj")[2],
  indiv = TRUE
)
}
\arguments{
\item{x}{character. 형태소 분석에 사용할 document.}

\item{type}{형태소 분석의 결과 유형.모든 품사, 명사, 동사 및 형용사와 같은
토큰화 결과 유형을 지정.
"morpheme", "noun", "noun2", "verb", "adj"중에서 선택. 기본값은 "noun"로
명사만 추출함.}

\item{indiv}{logical. 복수개의 문서일 때 개별 문서를 리스트로 반환할 지를 선택함.
TRUE이면 개별 리스트로 반환하고, FALSE이면 하나의 문자 벡터로 반환함.
기본값은 TRUE}
}
\value{
Hannnanum 형태소 분석기 결과 구조의 character 벡터.
}
\description{
KoNLP 패키지 내부에 포함된 한나눔 형태소 분석기의 호출
}
\examples{
\dontrun{
## Hannnanum 형태소 분석
morpho_hann("아버지가 방에 들어가신다.")
morpho_hann("아버지가 방에 들어가신다.", type = "verb")

morpho_hann(c("무궁화꽃이 피었습니다.", "나는 어제 올갱이국밥을 먹었다."))
morpho_hann(c("무궁화꽃이 피었습니다.", "나는 어제 올갱이국밥을 먹었다."), indiv = FALSE)
}
}
\seealso{
\code{\link{morpho_mecab}}
}
\author{
유충현
Maintainer: 유충현 <choonghyun.ryu@gmail.com>
}
