% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/president_speech.R
\docType{data}
\name{president_speech}
\alias{president_speech}
\title{대통령의 연설문}
\format{
6개의 변수와 2,408개의 관측치로 구성된 티블(tibble) 객체.:
\describe{
  \item{\code{president}}{character 연설 대통령}
  \item{\code{category}}{character 연설문 분야}
  \item{\code{type}}{character 연설문 유형}
  \item{\code{title}}{character 연설문 제목}
  \item{\code{speech_date}}{double 연설 일자}
  \item{\code{contents}}{character 연설문 내용}
}
}
\source{
"행정안전무 대통령기록관 홈페이지의 기록컬렉션>연설기록 페이지 <http://www.pa.go.kr/research/contents/speech/index.jsp>
}
\usage{
data(president_speech)
}
\description{
대통령기록연구실 홈페이지에서 수집한 역대 퇴임 대통령들의 연설문
}
\examples{
\dontrun{
data(president_speech)

head(president_speech)
}
}
\keyword{datasets}
